+++ t/html/pfappserver/config/bases/modules/networks_inline.js
